Moved write reports to data structures
