+++ normits_demand/core/__init__.py
@@ -12,6 +12,7 @@ from normits_demand.core.segments import get_segmentation_level
 from normits_demand.core.data_structures import multiply_and_aggregate_dvectors
 from normits_demand.core.data_structures import read_compressed_dvector
 from normits_demand.core.data_structures import from_pickle
+from normits_demand.core.data_structures import WriteReports
 
 # ## EXPOSE ERRORS ## #
 from normits_demand.core.zoning import ZoningError
