+++ normits_demand/core/data_structures.py
@@ -1459,4 +1459,45 @@ def from_pickle(path: nd.PathLike) -> Union[DVector, Any]:
     return obj
 
 
+class WriteReports:
 
+    def write_reports(self,
+                      dvec,
+                      segment_totals_path: nd.PathLike,
+                      ca_sector_path: nd.PathLike,
+                      ie_sector_path: nd.PathLike,
+                      ) -> None:
+        """
+        Writes segment, CA sector, and IE sector reports to disk
+
+        Parameters
+        ----------
+        dvec:
+            The Dvector to write the reports for
+
+        segment_totals_path:
+            Path to write the segment totals report to
+
+        ca_sector_path:
+            Path to write the CA sector report to
+
+        ie_sector_path:
+            Path to write the IE sector report to
+
+        Returns
+        -------
+        None
+        """
+        # Segment totals report
+        df = dvec.sum_zoning().to_df()
+        df.to_csv(segment_totals_path, index=False)
+
+        # Segment by CA Sector total reports
+        tfn_ca_sectors = nd.get_zoning_system('ca_sector_2020')
+        df = dvec.translate_zoning(tfn_ca_sectors)
+        df.to_df().to_csv(ca_sector_path, index=False)
+
+        # Segment by IE Sector total reports
+        ie_sectors = nd.get_zoning_system('ie_sector')
+        df = dvec.translate_zoning(ie_sectors).to_df()
+        df.to_csv(ie_sector_path, index=False)
