+++ normits_demand/core/definitions/segmentations/aggregate.csv
@@ -1,7 +1,6 @@
 in,out,common,translate
 full_tfntt_tfnat,full_tfntt,p;tfn_tt;m;tp,
 pure_demand,pure_demand_reporting,p;tfn_tt,tfn_at:tfn_agg_at;
-hb_notem_full_tfn,hb_tms_tp_week,p;m;tp,
 pure_attractions_ecat,pure_attractions,p;soc,
 full_attraction,p_m_soc,p;m;soc,
 hb_notem_output,hb_notem_output_no_tp,p;m;g;soc;ns;ca,
