--- normits_demand/core/definitions/segmentations/hb_full_tfn/naming_order.csv
@@ -1,7 +1,0 @@
-p
-m
-g
-soc
-ns
-ca
-tp
