+++ normits_demand/models/__init__.py
@@ -5,3 +5,4 @@ from normits_demand.models.notem_production_model import NHBProductionModel
 from normits_demand.models.notem_attraction_model import HBAttractionModel
 from normits_demand.models.notem_attraction_model import NHBAttractionModel
 from normits_demand.models.notem import NoTEM
+
