+++ normits_demand/pathing/__init__.py
@@ -13,4 +13,4 @@ from normits_demand.pathing.notem import HBProductionModelPaths
 from normits_demand.pathing.notem import HBAttractionModelPaths
 from normits_demand.pathing.notem import NHBProductionModelPaths
 from normits_demand.pathing.notem import NHBAttractionModelPaths
-from normits_demand.pathing.notem import WriteReports
+
